apu sorta works

+++ include/utils.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 namespace utils {
-    static uint8_t set_bit_value(uint8_t value, uint8_t bit_index, bool bit_value) {
+    static uint32_t set_bit_value(uint32_t value, uint32_t bit_index, bool bit_value) {
         return (value & ~(1 << bit_index)) | (bit_value << bit_index);
     }
 }
